requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
dj-database-url
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
